@@ -21,7 +21,6 @@ See COPYING file for the full LGPL text.
 Original ZIP package is copyrighted by Gilles Vollant and contributors,
 see quazip/(un)zip.h files for details. Basically it's the zlib license.
 */
-#ifndef __clang_analyzer__
 
 #include "quazipdir.h"
 #include "quazip_qt_compat.h"
@@ -34,25 +33,19 @@ class QuaZipDirPrivate: public QSharedData {
     friend class QuaZipDir;
 private:
     QuaZipDirPrivate(QuaZip *zip, const QString &dir = QString()):
-        zip(zip), dir(dir), caseSensitivity(QuaZip::csDefault),
-        filter(QDir::NoFilter), sorting(QDir::NoSort) {}
+        zip(zip), dir(dir) {}
     QuaZip *zip;
     QString dir;
-    QuaZip::CaseSensitivity caseSensitivity;
-    QDir::Filters filter;
+    QuaZip::CaseSensitivity caseSensitivity{QuaZip::csDefault};
+    QDir::Filters filter{QDir::NoFilter};
     QStringList nameFilters;
-    QDir::SortFlags sorting;
+    QDir::SortFlags sorting{QDir::NoSort};
     template<typename TFileInfoList>
     bool entryInfoList(QStringList nameFilters, QDir::Filters filter,
         QDir::SortFlags sort, TFileInfoList &result) const;
     inline QString simplePath() const {return QDir::cleanPath(dir);}
 };
 /// \endcond
-
-QuaZipDir::QuaZipDir(const QuaZipDir &that):
-    d(that.d)
-{
-}
 
 QuaZipDir::QuaZipDir(QuaZip *zip, const QString &dir):
     d(new QuaZipDirPrivate(zip, dir))
@@ -61,19 +54,13 @@ QuaZipDir::QuaZipDir(QuaZip *zip, const QString &dir):
         d->dir = d->dir.mid(1);
 }
 
-QuaZipDir::~QuaZipDir()
-{
-}
+QuaZipDir::QuaZipDir(const QuaZipDir &that) = default;
+QuaZipDir::~QuaZipDir() = default;
+QuaZipDir& QuaZipDir::operator=(const QuaZipDir &that) = default;
 
 bool QuaZipDir::operator==(const QuaZipDir &that)
 {
     return d->zip == that.d->zip && d->dir == that.d->dir;
-}
-
-QuaZipDir& QuaZipDir::operator=(const QuaZipDir &that)
-{
-    this->d = that.d;
-    return *this;
 }
 
 QString QuaZipDir::operator[](int pos) const
@@ -93,8 +80,9 @@ bool QuaZipDir::cd(const QString &directoryName)
         return true;
     }
     QString dirName = directoryName;
-    if (dirName.endsWith(QLatin1String("/")))
+    if (dirName.endsWith(QLatin1String("/"))) {
         dirName.chop(1);
+    }
     if (dirName.contains(QLatin1String("/"))) {
         QuaZipDir dir(*this);
         if (dirName.startsWith(QLatin1String("/"))) {
@@ -106,10 +94,7 @@ bool QuaZipDir::cd(const QString &directoryName)
                 return false;
         }
         QStringList path = dirName.split(QLatin1String("/"), SkipEmptyParts);
-        for (QStringList::const_iterator i = path.constBegin();
-                i != path.constEnd();
-                ++i) {
-            const QString &step = *i;
+        for (const auto& step : path) {
 #ifdef QUAZIP_QUAZIPDIR_DEBUG
             qDebug("QuaZipDir::cd(%s): going to %s",
                     dirName.toUtf8().constData(),
@@ -120,33 +105,28 @@ bool QuaZipDir::cd(const QString &directoryName)
         }
         d->dir = dir.path();
         return true;
-    } else { // no '/'
-        if (dirName == QLatin1String(".")) {
-            return true;
-        } else if (dirName == QLatin1String("..")) {
-            if (isRoot()) {
-                return false;
-            } else {
-                int slashPos = d->dir.lastIndexOf(QLatin1String("/"));
-                if (slashPos == -1) {
-                    d->dir = QLatin1String("");
-                } else {
-                    d->dir = d->dir.left(slashPos);
-                }
-                return true;
-            }
-        } else { // a simple subdirectory
-            if (exists(dirName)) {
-                if (isRoot())
-                    d->dir = dirName;
-                else
-                    d->dir += QLatin1String("/") + dirName;
-                return true;
-            } else {
-                return false;
-            }
-        }
     }
+    if (dirName == QLatin1String(".")) {
+        return true;
+    }
+    if (dirName == QLatin1String("..")) {
+        if (isRoot())
+            return false;
+        int slashPos = d->dir.lastIndexOf(QLatin1String("/"));
+        if (slashPos == -1)
+            d->dir = QLatin1String("");
+        else
+            d->dir = d->dir.left(slashPos);
+        return true;
+    }
+    if (exists(dirName)) {
+        if (isRoot())
+            d->dir = dirName;
+        else
+            d->dir += QLatin1String("/") + dirName;
+        return true;
+    }
+    return false;
 }
 
 bool QuaZipDir::cdUp()
@@ -197,10 +177,8 @@ static void QuaZipDir_convertInfoList(const QList<QuaZipFileInfo64> &from,
                                       QStringList &to)
 {
     to.clear();
-    for (QList<QuaZipFileInfo64>::const_iterator i = from.constBegin();
-            i != from.constEnd();
-            ++i) {
-        to.append(i->name);
+    for (const auto& file : from) {
+        to.append(file.name);
     }
 }
 
@@ -208,11 +186,9 @@ static void QuaZipDir_convertInfoList(const QList<QuaZipFileInfo64> &from,
                                       QList<QuaZipFileInfo> &to)
 {
     to.clear();
-    for (QList<QuaZipFileInfo64>::const_iterator i = from.constBegin();
-            i != from.constEnd();
-            ++i) {
+    for (const auto& file : from) {
         QuaZipFileInfo info32;
-        i->toQuaZipFileInfo(info32);
+        file.toQuaZipFileInfo(info32);
         to.append(info32);
     }
 }
@@ -249,12 +225,9 @@ class QuaZipDirComparator
 
 QString QuaZipDirComparator::getExtension(const QString &name)
 {
-    if (name.endsWith(QLatin1String(".")) || name.indexOf(QLatin1String("."), 1) == -1) {
-        return QLatin1String("");
-    } else {
+        if (name.endsWith(QLatin1String(".")) || name.indexOf(QLatin1String("."), 1) == -1)
+            return QLatin1String("");
         return name.mid(name.lastIndexOf(QLatin1String(".")) + 1);
-    }
-
 }
 
 int QuaZipDirComparator::compareStrings(const QString &string1,
@@ -263,13 +236,10 @@ int QuaZipDirComparator::compareStrings(const QString &string1,
     if (sort & QDir::LocaleAware) {
         if (sort & QDir::IgnoreCase) {
             return string1.toLower().localeAwareCompare(string2.toLower());
-        } else {
-            return string1.localeAwareCompare(string2);
         }
-    } else {
-        return string1.compare(string2, (sort & QDir::IgnoreCase)
-                ? Qt::CaseInsensitive : Qt::CaseSensitive);
+        return string1.localeAwareCompare(string2);
     }
+    return string1.compare(string2, (sort & QDir::IgnoreCase) ? Qt::CaseInsensitive : Qt::CaseSensitive);
 }
 
 bool QuaZipDirComparator::operator()(const QuaZipFileInfo64 &info1,
@@ -277,12 +247,13 @@ bool QuaZipDirComparator::operator()(const QuaZipFileInfo64 &info1,
 {
     QDir::SortFlags order = sort
         & (QDir::Name | QDir::Time | QDir::Size | QDir::Type);
-    if ((sort & QDir::DirsFirst) == QDir::DirsFirst
-            || (sort & QDir::DirsLast) == QDir::DirsLast) {
-        if (info1.name.endsWith(QLatin1String("/")) && !info2.name.endsWith(QLatin1String("/")))
+    if ((sort & QDir::DirsFirst) == QDir::DirsFirst || (sort & QDir::DirsLast) == QDir::DirsLast) {
+        if (info1.name.endsWith(QLatin1String("/")) && !info2.name.endsWith(QLatin1String("/"))) {
             return (sort & QDir::DirsFirst) == QDir::DirsFirst;
-        else if (!info1.name.endsWith(QLatin1String("/")) && info2.name.endsWith(QLatin1String("/")))
+        }
+        if (!info1.name.endsWith(QLatin1String("/")) && info2.name.endsWith(QLatin1String("/"))) {
             return (sort & QDir::DirsLast) == QDir::DirsLast;
+        }
     }
     bool result;
     int extDiff;
@@ -322,7 +293,7 @@ bool QuaZipDirComparator::operator()(const QuaZipFileInfo64 &info1,
 }
 
 template<typename TFileInfoList>
-bool QuaZipDirPrivate::entryInfoList(QStringList nameFilters, 
+bool QuaZipDirPrivate::entryInfoList(QStringList nameFilters,
     QDir::Filters filter, QDir::SortFlags sort, TFileInfoList &result) const
 {
     QString basePath = simplePath();
@@ -404,10 +375,9 @@ QList<QuaZipFileInfo> QuaZipDir::entryInfoList(const QStringList &nameFilters,
     QDir::Filters filters, QDir::SortFlags sort) const
 {
     QList<QuaZipFileInfo> result;
-    if (d->entryInfoList(nameFilters, filters, sort, result))
-        return result;
-    else
+    if (!d->entryInfoList(nameFilters, filters, sort, result))
         return QList<QuaZipFileInfo>();
+    return result;
 }
 
 QList<QuaZipFileInfo> QuaZipDir::entryInfoList(QDir::Filters filters,
@@ -420,10 +390,9 @@ QList<QuaZipFileInfo64> QuaZipDir::entryInfoList64(const QStringList &nameFilter
     QDir::Filters filters, QDir::SortFlags sort) const
 {
     QList<QuaZipFileInfo64> result;
-    if (d->entryInfoList(nameFilters, filters, sort, result))
-        return result;
-    else
+    if (!d->entryInfoList(nameFilters, filters, sort, result))
         return QList<QuaZipFileInfo64>();
+    return result;
 }
 
 QList<QuaZipFileInfo64> QuaZipDir::entryInfoList64(QDir::Filters filters,
@@ -436,10 +405,9 @@ QStringList QuaZipDir::entryList(const QStringList &nameFilters,
     QDir::Filters filters, QDir::SortFlags sort) const
 {
     QStringList result;
-    if (d->entryInfoList(nameFilters, filters, sort, result))
-        return result;
-    else
+    if (!d->entryInfoList(nameFilters, filters, sort, result))
         return QStringList();
+    return result;
 }
 
 QStringList QuaZipDir::entryList(QDir::Filters filters,
@@ -465,33 +433,29 @@ bool QuaZipDir::exists(const QString &filePath) const
 #endif
         QuaZipDir dir(*this);
         return dir.cd(fileInfo.path()) && dir.exists(fileInfo.fileName());
-    } else {
-        if (fileName == QLatin1String("..")) {
-            return !isRoot();
-        } else if (fileName == QLatin1String(".")) {
-            return true;
-        } else {
-            QStringList entries = entryList(QDir::AllEntries, QDir::NoSort);
+    }
+    if (fileName == QLatin1String("..")) {
+        return !isRoot();
+    }
+    if (fileName == QLatin1String(".")) {
+        return true;
+    }
+    QStringList entries = entryList(QDir::AllEntries, QDir::NoSort);
 #ifdef QUAZIP_QUAZIPDIR_DEBUG
             qDebug("QuaZipDir::exists(): looking for %s",
                     fileName.toUtf8().constData());
-            for (QStringList::const_iterator i = entries.constBegin();
-                    i != entries.constEnd();
-                    ++i) {
+            for (const auto& entry : entries) {
                 qDebug("QuaZipDir::exists(): entry: %s",
-                        i->toUtf8().constData());
+                        entry.toUtf8().constData());
             }
 #endif
             Qt::CaseSensitivity cs = QuaZip::convertCaseSensitivity(
                     d->caseSensitivity);
             if (filePath.endsWith(QLatin1String("/"))) {
                 return entries.contains(filePath, cs);
-            } else {
-                return entries.contains(fileName, cs)
-                    || entries.contains(fileName + QLatin1String("/"), cs);
             }
-        }
-    }
+            return entries.contains(fileName, cs) ||
+                   entries.contains(fileName + QLatin1String("/"), cs);
 }
 
 bool QuaZipDir::exists() const
@@ -567,5 +531,3 @@ QDir::SortFlags QuaZipDir::sorting() const
 {
     return d->sorting;
 }
-
-#endif
