@@ -32,51 +32,67 @@
 #define SEEK_SET    0
 #endif
 
-voidpf call_zopen64 (const zlib_filefunc64_32_def* pfilefunc,voidpf file,int mode)
+voidpf call_zopen64(const zlib_filefunc64_32_def* pfilefunc, voidpf file, int mode)
 {
-    if (pfilefunc->zfile_func64.zopen64_file != nullptr)
-        return (*(pfilefunc->zfile_func64.zopen64_file)) (pfilefunc->zfile_func64.opaque,file,mode);
-    else
-    {
-        return (*(pfilefunc->zopen32_file))(pfilefunc->zfile_func64.opaque,file,mode);
-    }
+    auto func = pfilefunc->zfile_func64.zopen64_file != nullptr
+        ? pfilefunc->zfile_func64.zopen64_file
+        : pfilefunc->zopen32_file;
+    return (*func)(pfilefunc->zfile_func64.opaque, file, mode);
 }
 
-int call_zseek64 (const zlib_filefunc64_32_def* pfilefunc,voidpf filestream, ZPOS64_T offset, int origin)
+namespace {
+
+int zseek64(const zlib_filefunc64_32_def* pfilefunc, voidpf filestream, ZPOS64_T offset, int origin)
 {
-    if (pfilefunc->zfile_func64.zseek64_file != nullptr)
-        return (*(pfilefunc->zfile_func64.zseek64_file)) (pfilefunc->zfile_func64.opaque,filestream,offset,origin);
-    else
-    {
-        uLong offsetTruncated = (uLong)offset;
-        if (offsetTruncated != offset)
-            return -1;
-        else
-            return (*(pfilefunc->zseek32_file))(pfilefunc->zfile_func64.opaque,filestream,offsetTruncated,origin);
-    }
+    return (*pfilefunc->zfile_func64.zseek64_file)(pfilefunc->zfile_func64.opaque, filestream, offset, origin);
 }
 
-ZPOS64_T call_ztell64 (const zlib_filefunc64_32_def* pfilefunc,voidpf filestream)
+
+int zseek32(const zlib_filefunc64_32_def* pfilefunc, voidpf filestream, ZPOS64_T offset, int origin)
 {
-    if (pfilefunc->zfile_func64.zseek64_file != nullptr)
-        return (*(pfilefunc->zfile_func64.ztell64_file)) (pfilefunc->zfile_func64.opaque,filestream);
-    else
-    {
-        uLong tell_uLong = (*(pfilefunc->ztell32_file))(pfilefunc->zfile_func64.opaque,filestream);
-        if ((tell_uLong) == ((uLong)-1))
-            return (ZPOS64_T)-1;
-        else
-            return tell_uLong;
-    }
+    uLong offsetTruncated = static_cast<uLong>(offset);
+    if (offsetTruncated != offset)
+        return -1;
+    return (*pfilefunc->zseek32_file)(pfilefunc->zfile_func64.opaque, filestream, offsetTruncated, origin);
+}
+
+}
+
+int call_zseek64(const zlib_filefunc64_32_def* pfilefunc, voidpf filestream, ZPOS64_T offset, int origin)
+{
+    auto func = pfilefunc->zfile_func64.zseek64_file != nullptr ? zseek64 : zseek32;
+    return (*func)(pfilefunc, filestream, offset, origin);
+
+}
+
+namespace {
+
+ZPOS64_T ztell64(const zlib_filefunc64_32_def* pfilefunc, voidpf filestream)
+{
+    return (*pfilefunc->zfile_func64.ztell64_file) (pfilefunc->zfile_func64.opaque, filestream);
+}
+
+ZPOS64_T ztell32(const zlib_filefunc64_32_def* pfilefunc, voidpf filestream)
+{
+    uLong tell_uLong = (*pfilefunc->ztell32_file)(pfilefunc->zfile_func64.opaque, filestream);
+    if (tell_uLong == static_cast<uLong>(-1))
+        return static_cast<ZPOS64_T>(-1);
+    return tell_uLong;
+}
+
+}
+
+ZPOS64_T call_ztell64(const zlib_filefunc64_32_def* pfilefunc, voidpf filestream)
+{
+    auto *func = pfilefunc->zfile_func64.zseek64_file != nullptr ? ztell64 : ztell32;
+    return (*func)(pfilefunc, filestream);
+
 }
 
 /// @cond internal
 struct QIODevice_descriptor {
     // Position only used for writing to sequential devices.
-    qint64 pos;
-    inline QIODevice_descriptor():
-        pos(0)
-    {}
+    qint64 pos{0};
 };
 /// @endcond
 
@@ -94,28 +110,27 @@ voidpf ZCALLBACK qiodevice_open_file_func (
         desiredMode = QIODevice::ReadWrite;
     else if (mode & ZLIB_FILEFUNC_MODE_CREATE)
         desiredMode = QIODevice::WriteOnly;
+    else
+        return nullptr;
     if (iodevice->isOpen()) {
-        if ((iodevice->openMode() & desiredMode) == desiredMode) {
-            if (desiredMode != QIODevice::WriteOnly
-                    && iodevice->isSequential()) {
-                // We can use sequential devices only for writing.
-                delete d;
-                return nullptr;
-            } else {
-                if ((desiredMode & QIODevice::WriteOnly) != 0) {
-                    // open for writing, need to seek existing device
-                    if (!iodevice->isSequential()) {
-                        iodevice->seek(0);
-                    } else {
-                        d->pos = iodevice->pos();
-                    }
-                }
-            }
-            return iodevice;
-        } else {
+        if ((iodevice->openMode() & desiredMode) != desiredMode) {
             delete d;
             return nullptr;
         }
+        if (desiredMode != QIODevice::WriteOnly && iodevice->isSequential()) {
+            // We can use sequential devices only for writing.
+            delete d;
+            return nullptr;
+        }
+        if ((desiredMode & QIODevice::WriteOnly) != 0) {
+            // open for writing, need to seek existing device
+            if (!iodevice->isSequential()) {
+              iodevice->seek(0);
+            } else {
+              d->pos = iodevice->pos();
+            }
+        }
+        return iodevice;
     }
     iodevice->open(desiredMode);
     if (iodevice->isOpen()) {
@@ -124,13 +139,11 @@ voidpf ZCALLBACK qiodevice_open_file_func (
             iodevice->close();
             delete d;
             return nullptr;
-        } else {
-            return iodevice;
         }
-    } else {
-        delete d;
-        return nullptr;
+        return iodevice;
     }
+    delete d;
+    return nullptr;
 }
 
 
@@ -142,9 +155,9 @@ uLong ZCALLBACK qiodevice_read_file_func (
 {
     QIODevice_descriptor *d = reinterpret_cast<QIODevice_descriptor*>(opaque);
     QIODevice *iodevice = reinterpret_cast<QIODevice*>(stream);
-    qint64 ret64 = iodevice->read((char*)buf,size);
+    qint64 ret64 = iodevice->read(static_cast<char*>(buf),size);
     uLong ret;
-    ret = (uLong) ret64;
+    ret = static_cast<uLong>(ret64);
     if (ret64 != -1) {
         d->pos += ret64;
     }
@@ -165,7 +178,7 @@ uLong ZCALLBACK qiodevice_write_file_func (
     if (ret64 != -1) {
         d->pos += ret64;
     }
-    ret = (uLong) ret64;
+    ret = static_cast<uLong>(ret64);
     return ret;
 }
 
@@ -209,24 +222,22 @@ int ZCALLBACK qiodevice_seek_file_func (
 {
     QIODevice *iodevice = reinterpret_cast<QIODevice*>(stream);
     if (iodevice->isSequential()) {
-        if (origin == ZLIB_FILEFUNC_SEEK_END
-                && offset == 0) {
+        if (origin == ZLIB_FILEFUNC_SEEK_END && offset == 0) {
             // sequential devices are always at end (needed in mdAppend)
             return 0;
-        } else {
-            qWarning("qiodevice_seek_file_func() called for sequential device");
-            return -1;
         }
+        qWarning("qiodevice_seek_file_func() called for sequential device");
+        return -1;
     }
     uLong qiodevice_seek_result=0;
     int ret;
     switch (origin)
     {
     case ZLIB_FILEFUNC_SEEK_CUR :
-        qiodevice_seek_result = ((QIODevice*)stream)->pos() + offset;
+        qiodevice_seek_result = (static_cast<QIODevice*>(stream))->pos() + offset;
         break;
     case ZLIB_FILEFUNC_SEEK_END :
-        qiodevice_seek_result = ((QIODevice*)stream)->size() - offset;
+        qiodevice_seek_result = (static_cast<QIODevice*>(stream))->size() - offset;
         break;
     case ZLIB_FILEFUNC_SEEK_SET :
         qiodevice_seek_result = offset;
@@ -246,24 +257,22 @@ int ZCALLBACK qiodevice64_seek_file_func (
 {
     QIODevice *iodevice = reinterpret_cast<QIODevice*>(stream);
     if (iodevice->isSequential()) {
-        if (origin == ZLIB_FILEFUNC_SEEK_END
-                && offset == 0) {
+        if (origin == ZLIB_FILEFUNC_SEEK_END && offset == 0) {
             // sequential devices are always at end (needed in mdAppend)
             return 0;
-        } else {
-            qWarning("qiodevice_seek_file_func() called for sequential device");
-            return -1;
         }
+        qWarning("qiodevice_seek_file_func() called for sequential device");
+        return -1;
     }
     qint64 qiodevice_seek_result=0;
     int ret;
     switch (origin)
     {
     case ZLIB_FILEFUNC_SEEK_CUR :
-        qiodevice_seek_result = ((QIODevice*)stream)->pos() + offset;
+        qiodevice_seek_result = (static_cast<QIODevice*>(stream))->pos() + offset;
         break;
     case ZLIB_FILEFUNC_SEEK_END :
-        qiodevice_seek_result = ((QIODevice*)stream)->size() - offset;
+        qiodevice_seek_result = (static_cast<QIODevice*>(stream))->size() - offset;
         break;
     case ZLIB_FILEFUNC_SEEK_SET :
         qiodevice_seek_result = offset;
