@@ -34,7 +34,7 @@ QuaAdler32::QuaAdler32()
 
 quint32 QuaAdler32::calculate(const QByteArray &data)
 {
-	return adler32( adler32(0L, Z_NULL, 0), (const Bytef*)data.data(), data.size() );
+	return adler32( adler32(0L, Z_NULL, 0), reinterpret_cast<const Bytef*>(data.data()), data.size() );
 }
 
 void QuaAdler32::reset()
@@ -44,7 +44,7 @@ void QuaAdler32::reset()
 
 void QuaAdler32::update(const QByteArray &buf)
 {
-	checksum = adler32( checksum, (const Bytef*)buf.data(), buf.size() );
+	checksum = adler32( checksum, reinterpret_cast<const Bytef*>(buf.data()), buf.size() );
 }
 
 quint32 QuaAdler32::value()
