@@ -1,5 +1,3 @@
 #include "quachecksum32.h"
 
-QuaChecksum32::~QuaChecksum32()
-{
-}
+QuaChecksum32::~QuaChecksum32() = default;
