@@ -33,7 +33,7 @@ QuaCrc32::QuaCrc32()
 
 quint32 QuaCrc32::calculate(const QByteArray &data)
 {
-	return crc32( crc32(0L, Z_NULL, 0), (const Bytef*)data.data(), data.size() );
+	return crc32( crc32(0L, Z_NULL, 0), reinterpret_cast<const Bytef*>(data.data()), data.size() );
 }
 
 void QuaCrc32::reset()
@@ -43,7 +43,7 @@ void QuaCrc32::reset()
 
 void QuaCrc32::update(const QByteArray &buf)
 {
-	checksum = crc32( checksum, (const Bytef*)buf.data(), buf.size() );
+	checksum = crc32( checksum, reinterpret_cast<const Bytef*>(buf.data()), buf.size() );
 }
 
 quint32 QuaCrc32::value()
