@@ -66,7 +66,8 @@ bool QuaGzipFilePrivate::open(FileId id, QIODevice::OpenMode mode,
         error = QuaGzipFile::tr("Opening gzip for both reading"
             " and writing is not supported");
         return false;
-    } else if ((mode & QIODevice::ReadOnly) != 0) {
+    }
+    if ((mode & QIODevice::ReadOnly) != 0) {
         modeString[0] = 'r';
     } else if ((mode & QIODevice::WriteOnly) != 0) {
         modeString[0] = 'w';
@@ -157,16 +158,15 @@ void QuaGzipFile::close()
 
 qint64 QuaGzipFile::readData(char *data, qint64 maxSize)
 {
-    return gzread(d->gzd, (voidp)data, (unsigned)maxSize);
+    return gzread(d->gzd, (voidp)data, static_cast<unsigned>(maxSize));
 }
 
 qint64 QuaGzipFile::writeData(const char *data, qint64 maxSize)
 {
     if (maxSize == 0)
         return 0;
-    int written = gzwrite(d->gzd, (voidp)data, (unsigned)maxSize);
+    int written = gzwrite(d->gzd, (voidp)data, static_cast<unsigned>(maxSize));
     if (written == 0)
         return -1;
-    else
-        return written;
+    return written;
 }
