@@ -36,7 +36,15 @@ see quazip/(un)zip.h files for details. Basically it's the zlib license.
 #ifdef QUAZIP_STATIC
 #define QUAZIP_EXPORT
 #else
-#include <quazip_export.h>
+/**
+ * When building a DLL with MSVC, QUAZIP_BUILD must be defined.
+ * qglobal.h takes care of defining Q_DECL_* correctly for msvc/gcc.
+ */
+#if defined(QUAZIP_BUILD)
+	#define QUAZIP_EXPORT Q_DECL_EXPORT
+#else
+	#define QUAZIP_EXPORT Q_DECL_IMPORT
+#endif
 #endif // QUAZIP_STATIC
 
 #ifdef __GNUC__
