@@ -285,7 +285,7 @@ bool QuaZipFile::open(OpenMode mode, int *method, int *level, bool raw, const ch
       }
       if(p->zip->getMode()!=QuaZip::mdUnzip) {
         qWarning("QuaZipFile::open(): file open mode %d incompatible with ZIP open mode %d",
-            (int)mode, (int)p->zip->getMode());
+            (int)mode, static_cast<int>(p->zip->getMode()));
         return false;
       }
       if(!p->zip->hasCurrentFile()) {
@@ -293,13 +293,13 @@ bool QuaZipFile::open(OpenMode mode, int *method, int *level, bool raw, const ch
         return false;
       }
     }
-    p->setZipError(unzOpenCurrentFile3(p->zip->getUnzFile(), method, level, (int)raw, password));
-    if(p->zipError==UNZ_OK) {
-      setOpenMode(mode);
-      p->raw=raw;
-      return true;
-    } else
+    p->setZipError(unzOpenCurrentFile3(p->zip->getUnzFile(), method, level, static_cast<int>(raw), password));
+    if(p->zipError != UNZ_OK) {
       return false;
+    }
+    setOpenMode(mode);
+    p->raw=raw;
+    return true;
   }
   qWarning("QuaZipFile::open(): open mode %d not supported by this function", (int)mode);
   return false;
@@ -327,7 +327,7 @@ bool QuaZipFile::open(OpenMode mode, const QuaZipNewInfo& info,
     }
     if(p->zip->getMode()!=QuaZip::mdCreate&&p->zip->getMode()!=QuaZip::mdAppend&&p->zip->getMode()!=QuaZip::mdAdd) {
       qWarning("QuaZipFile::open(): file open mode %d incompatible with ZIP open mode %d",
-          (int)mode, (int)p->zip->getMode());
+          (int)mode, static_cast<int>(p->zip->getMode()));
       return false;
     }
     info_z.tmz_date.tm_year=info.dateTime.date().year();
@@ -337,8 +337,8 @@ bool QuaZipFile::open(OpenMode mode, const QuaZipNewInfo& info,
     info_z.tmz_date.tm_min=info.dateTime.time().minute();
     info_z.tmz_date.tm_sec=info.dateTime.time().second();
     info_z.dosDate = 0;
-    info_z.internal_fa=(uLong)info.internalAttr;
-    info_z.external_fa=(uLong)info.externalAttr;
+    info_z.internal_fa=static_cast<uLong>(info.internalAttr);
+    info_z.external_fa=static_cast<uLong>(info.externalAttr);
     if (p->zip->isDataDescriptorWritingEnabled())
         zipSetFlags(p->zip->getZipFile(), ZIP_WRITE_DATA_DESCRIPTOR);
     else
@@ -353,23 +353,23 @@ bool QuaZipFile::open(OpenMode mode, const QuaZipNewInfo& info,
           p->zip->isUtf8Enabled()
             ? info.comment.toUtf8().constData()
             : p->zip->getCommentCodec()->fromUnicode(info.comment).constData(),
-          method, level, (int)raw,
+          method, level, static_cast<int>(raw),
           windowBits, memLevel, strategy,
-          password, (uLong)crc,
+          password, static_cast<uLong>(crc),
           (p->zip->getOsCode() << 8) | QUAZIP_VERSION_MADE_BY,
           0,
           p->zip->isZip64Enabled()));
-    if(p->zipError==UNZ_OK) {
-      p->writePos=0;
-      setOpenMode(mode);
-      p->raw=raw;
-      if(raw) {
-        p->crc=crc;
-        p->uncompressedSize=info.uncompressedSize;
-      }
-      return true;
-    } else
+    if (p->zipError != UNZ_OK) {
       return false;
+    }
+    p->writePos=0;
+    setOpenMode(mode);
+    p->raw=raw;
+    if(raw) {
+      p->crc=crc;
+      p->uncompressedSize=info.uncompressedSize;
+    }
+    return true;
   }
   qWarning("QuaZipFile::open(): open mode %d not supported by this function", (int)mode);
   return false;
@@ -395,8 +395,7 @@ qint64 QuaZipFile::pos()const
       // but thankfully bytesAvailable() returns the number of
       // bytes buffered, so we know how far ahead we are.
     return unztell64(p->zip->getUnzFile()) - QIODevice::bytesAvailable();
-  else
-    return p->writePos;
+  return p->writePos;
 }
 
 bool QuaZipFile::atEnd()const
@@ -413,8 +412,7 @@ bool QuaZipFile::atEnd()const
       // the same problem as with pos()
     return QIODevice::bytesAvailable() == 0
         && unzeof(p->zip->getUnzFile())==1;
-  else
-    return true;
+  return true;
 }
 
 qint64 QuaZipFile::size()const
@@ -425,8 +423,7 @@ qint64 QuaZipFile::size()const
   }
   if(openMode()&ReadOnly)
     return p->raw?csize():usize();
-  else
-    return p->writePos;
+  return p->writePos;
 }
 
 qint64 QuaZipFile::csize()const
@@ -454,12 +451,11 @@ qint64 QuaZipFile::usize()const
 bool QuaZipFile::getFileInfo(QuaZipFileInfo *info)
 {
     QuaZipFileInfo64 info64;
-    if (getFileInfo(&info64)) {
-        info64.toQuaZipFileInfo(*info);
-        return true;
-    } else {
+    if (!getFileInfo(&info64)) {
         return false;
     }
+    info64.toQuaZipFileInfo(*info);
+    return true;
 }
 
 bool QuaZipFile::getFileInfo(QuaZipFileInfo64 *info)
@@ -498,9 +494,9 @@ void QuaZipFile::close()
 qint64 QuaZipFile::readData(char *data, qint64 maxSize)
 {
   p->setZipError(UNZ_OK);
-  qint64 bytesRead=unzReadCurrentFile(p->zip->getUnzFile(), data, (unsigned)maxSize);
+  qint64 bytesRead=unzReadCurrentFile(p->zip->getUnzFile(), data, static_cast<unsigned>(maxSize));
   if (bytesRead < 0) {
-    p->setZipError((int) bytesRead);
+    p->setZipError(static_cast<int>(bytesRead));
     return -1;
   }
   return bytesRead;
@@ -509,12 +505,12 @@ qint64 QuaZipFile::readData(char *data, qint64 maxSize)
 qint64 QuaZipFile::writeData(const char* data, qint64 maxSize)
 {
   p->setZipError(ZIP_OK);
-  p->setZipError(zipWriteInFileInZip(p->zip->getZipFile(), data, (uint)maxSize));
-  if(p->zipError!=ZIP_OK) return -1;
-  else {
-    p->writePos+=maxSize;
-    return maxSize;
+  p->setZipError(zipWriteInFileInZip(p->zip->getZipFile(), data, static_cast<uint>(maxSize)));
+  if (p->zipError != ZIP_OK) {
+    return -1;
   }
+  p->writePos += maxSize;
+  return maxSize;
 }
 
 QString QuaZipFile::getFileName() const
