@@ -5,9 +5,8 @@
 
 class TestQuaZipNewInfo : public QObject
 {
+using QObject::QObject;
     Q_OBJECT
-public:
-    explicit TestQuaZipNewInfo(QObject *parent = 0);
 private slots:
     void setFileNTFSTimes();
 };
